Extract service from controller
